+++ src/openassetio-python/tests/package/pluginSystem/resources/entryPoint/src/packaged_plugin/PackagePlugin.py
@@ -1,0 +1,13 @@
+"""
+Provides a test PythonPluginSystemPlugin implementation.
+"""
+
+from openassetio.pluginSystem import PythonPluginSystemPlugin
+
+
+class PackagePlugin(PythonPluginSystemPlugin):
+    # pylint: disable=missing-class-docstring
+
+    @classmethod
+    def identifier(cls):
+        return "org.openassetio.test.pluginSystem.resources.packagePlugin"